refactor(editor): scope the durations move to the debounce unit only
Per Trigger's call on N8N-25: keep only the debounce unit (getDebounceTime +
DEBOUNCE_TIME) in @n8n/composables, colocated with its consumer useDebounce.
getDebounceTime reads sessionStorage (browser-only) and is a function, so it
must not land in the backend-imported @n8n/constants. The general timing
constants (TIME, *_IN_MILLIS, throttle durations, …) stay in editor-ui's
app/constants for the Wave 0.3 constants split; @/app/constants/durations
re-exports the debounce unit so existing call sites keep working. CAT-3686.
