+++ packages/frontend/@n8n/composables/src/durations.ts
@@ -1,30 +1,11 @@
 /**
- * Performance Optimizations
+ * Debounce timing unit for `useDebounce`, colocated with its only consumer.
+ *
+ * `getDebounceTime` reads a `sessionStorage` multiplier (browser-only) and is a
+ * function, so it must not live in the backend-imported `@n8n/constants`. The
+ * general timing constants (`TIME`, `*_IN_MILLIS`, throttle durations, …) stay
+ * in editor-ui's `app/constants` for the Wave 0.3 constants split (CAT-3686).
  */
-
-export const LOGS_EXECUTION_DATA_THROTTLE_DURATION = 1000;
-export const CANVAS_EXECUTION_DATA_THROTTLE_DURATION = 500;
-
-export const EXPRESSION_EDITOR_PARSER_TIMEOUT = 15_000; // ms
-
-export const CLOUD_TRIAL_CHECK_INTERVAL = 5000;
-
-/**
- * Units of time in milliseconds
- */
-
-export const TIME = {
-	SECOND: 1000,
-	MINUTE: 60 * 1000,
-	HOUR: 60 * 60 * 1000,
-	DAY: 24 * 60 * 60 * 1000,
-};
-
-export const THREE_DAYS_IN_MILLIS = 3 * TIME.DAY;
-export const SEVEN_DAYS_IN_MILLIS = 7 * TIME.DAY;
-export const SIX_MONTHS_IN_MILLIS = 6 * 30 * TIME.DAY;
-
-export const LOADING_ANIMATION_MIN_DURATION = 1000;
 
 /**
  * Get debounce time with optional multiplier from sessionStorage.
